+++ pkg/apiserver/metrics/prom_resolve.go
@@ -6,8 +6,6 @@ import (
 	"bytes"
 	"encoding/json"
 	"fmt"
-	"github.com/pingcap/log"
-	"go.uber.org/zap"
 	"net"
 	"net/url"
 	"strconv"
@@ -73,7 +71,6 @@ func (s *Service) resolveCustomizedPromAddress(acceptInvalidAddr bool) (string, 
 		if err != nil {
 			return "", err
 		}
-		log.Warn("get prometheus address from PD config", zap.String("addr", addr))
 		return addr, nil
 	}
 	return "", nil
@@ -108,7 +105,6 @@ func (s *Service) resolveFinalPromAddress() (string, error) {
 		return "", err
 	}
 	if addr != "" {
-		log.Warn("get prometheus address from topology", zap.String("addr", addr))
 		return addr, nil
 	}
 	return "", nil
