+++ pkg/apiserver/metrics/service.go
@@ -4,6 +4,9 @@ package metrics
 
 import (
 	"context"
+	"crypto/tls"
+	"net"
+	"net/http"
 	"time"
 
 	"github.com/joomcode/errorx"
@@ -40,10 +43,24 @@ type Service struct {
 
 	promRequestGroup singleflight.Group
 	promAddressCache atomic.Value
+	httpClient       *http.Client
 }
 
 func NewService(lc fx.Lifecycle, p ServiceParams) *Service {
-	s := &Service{params: p}
+	tlsConfig := &tls.Config{InsecureSkipVerify: true}
+	s := &Service{
+		params: p,
+		httpClient: &http.Client{
+			Transport: &http.Transport{
+				DialTLS: func(network, addr string) (net.Conn, error) {
+					conn, err := tls.Dial(network, addr, tlsConfig)
+					return conn, err
+				},
+				TLSClientConfig: tlsConfig,
+			},
+			Timeout: defaultPromQueryTimeout,
+		},
+	}
 
 	lc.Append(fx.Hook{
 		OnStart: func(ctx context.Context) error {
